Move SimulationResults to separate file so it can be loaded on platforms without QT5 such as ARM Graviton64
