+++ wmpl/MetSim/AutoRefineFit.py
@@ -11,7 +11,8 @@ import scipy.optimize
 
 from wmpl.Formats.Met import loadMet
 from wmpl.Trajectory.Trajectory import Trajectory
-from wmpl.MetSim.GUI import loadConstants, saveConstants, SimulationResults, MetObservations
+from wmpl.MetSim.GUI import loadConstants, saveConstants, MetObservations
+from wmpl.MetSim.SimResults import SimulationResults
 from wmpl.MetSim.MetSimErosion import runSimulation, Constants
 from wmpl.Utils.AtmosphereDensity import fitAtmPoly
 from wmpl.Utils.Math import meanAngle
